backend/pyproject.toml
[project]
name = "dechord-api"
version = "0.1.0"
description = "DeChord web API - chord/key/tempo detection from audio"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]",
    "python-multipart",
    "python-dotenv",
    "madmom",
    "libsql-client>=0.3.1",
    "demucs; python_version < '3.14'",
    "lameenc; python_version < '3.14'",
    "mido>=1.3.3",
    "pyguitarpro>=0.10.1",
]

[tool.uv.sources]
madmom = { git = "https://github.com/CPJKU/madmom" }

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
